changes on location module

+++ resources/views/admin/location/index.blade.php
@@ -9,8 +9,12 @@
     <div class="animate__animated animate__fadeInUp">
         <div class="msg-content white-bg radius-50 space-30 d-flex align-items-center">
             <h2 class="mb-md-0">@lang('cruds.location.title_singular')</h2>
-            <a href="javascript:void(0)" class="add_btn dash-btn green-bg w-115" id="addLocationBtn" >+ @lang('global.add') @lang('global.new')</a><div class="d-sm-none w-100"></div>
-            <button class="del_btn dash-btn red-bg w-115 me-md-1" id="deleteAllLocation">@lang('global.delete')</button>
+            @can('location_create')
+                <a href="javascript:void(0)" class="add_btn dash-btn green-bg w-115" id="addLocationBtn" >+ @lang('global.add') @lang('global.new')</a><div class="d-sm-none w-100"></div>
+            @endcan
+            @can('location_delete')
+                <button class="del_btn dash-btn red-bg w-115 me-md-1" id="deleteAllLocation">@lang('global.delete')</button>
+            @endcan
         </div>
         <div class="c-admin position-relative">
             <div class="table-responsive">
@@ -26,216 +30,221 @@
 {!! $dataTable->scripts() !!}
 
 <script>
-    // Add Location Modal
-    $(document).on('click', '#addLocationBtn', function(e){
-        e.preventDefault();
-        // $('#pageloader').css('display', 'block');
-        $.ajax({
-            type: 'get',
-            url: "{{ route('locations.create') }}",
-            dataType: 'json',
-            success: function (response) {
-                // $('#pageloader').css('display', 'none');
-                if(response.success) {
-                    $('.popup_render_div').html(response.htmlView);
-                    $('#addLocationModal').modal('show');
+    @can('location_create')
+        // Add Location Modal
+        $(document).on('click', '#addLocationBtn', function(e){
+            e.preventDefault();
+            // $('#pageloader').css('display', 'block');
+            $.ajax({
+                type: 'get',
+                url: "{{ route('locations.create') }}",
+                dataType: 'json',
+                success: function (response) {
+                    // $('#pageloader').css('display', 'none');
+                    if(response.success) {
+                        $('.popup_render_div').html(response.htmlView);
+                        $('#addLocationModal').modal('show');
+                    }
                 }
-            }
-        })
-    });
+            })
+        });
 
-    // Submit Add Location Form
-    $(document).on('submit', '#addLocationForm', function (e) {
-        e.preventDefault();
-        $('.validation-error-block').remove();
-        $(".submitBtn").attr('disabled', true);
+        // Submit Add Location Form
+        $(document).on('submit', '#addLocationForm', function (e) {
+            e.preventDefault();
+            $('.validation-error-block').remove();
+            $(".submitBtn").attr('disabled', true);
 
-        var formData = new FormData(this);
+            var formData = new FormData(this);
 
-        $.ajax({
-            type: 'post',
-            url: "{{route('locations.store')}}",
-            dataType: 'json',
-            contentType: false,
-            processData: false,
-            data: formData,
-            success: function (response) {                
-                if(response.success) {
-                    $('#addLocationModal').modal('hide');
-                    fireSuccessSwal('Success',response.message);
-                    $('#location-table').DataTable().ajax.reload(null, false);
+            $.ajax({
+                type: 'post',
+                url: "{{route('locations.store')}}",
+                dataType: 'json',
+                contentType: false,
+                processData: false,
+                data: formData,
+                success: function (response) {                
+                    if(response.success) {
+                        $('#addLocationModal').modal('hide');
+                        fireSuccessSwal('Success',response.message);
+                        $('#location-table').DataTable().ajax.reload(null, false);
+                    }
+                },
+                error: function (response) {
+                    if(response.error_type == 'something_error'){
+                        fireErrorSwal('Error',response.message);
+                    } else {                    
+                        var errorLabelTitle = '';
+                        $.each(response.responseJSON.errors, function (key, item) {
+                            errorLabelTitle = '<span class="validation-error-block">'+item+'</sapn>';
+                            if(key == 'resume' || key == 'gender'){
+                                $('#'+key).html(errorLabelTitle);
+                            }
+                            else{
+                                $(errorLabelTitle).insertAfter("input[name='"+key+"']");
+                            }                    
+                        });
+                    }
+                },
+                complete: function(res){
+                    $(".submitBtn").attr('disabled', false);
                 }
-            },
-            error: function (response) {
-                if(response.error_type == 'something_error'){
-                    fireErrorSwal('Error',response.message);
-                } else {                    
-                    var errorLabelTitle = '';
-                    $.each(response.responseJSON.errors, function (key, item) {
-                        errorLabelTitle = '<span class="validation-error-block">'+item+'</sapn>';
-                        if(key == 'resume' || key == 'gender'){
-                            $('#'+key).html(errorLabelTitle);
+            });                    
+        });
+    @endcan
+
+    @can('location_edit')
+        // Edit Location Modal
+        $(document).on("click",".editLocationBtn", function() {
+            // $('#pageloader').css('display', 'flex');
+            var url = $(this).data('href');
+            $.ajax({
+                type: 'get',
+                url: url,
+                dataType: 'json',
+                //data: formData,
+                success: function (response) {
+                    // $('#pageloader').css('display', 'none');
+                    if(response.success) {
+                        $('.popup_render_div').html(response.htmlView);
+                        $('#editLocationModal').modal('show');
+                    }
+                }
+            });
+        });
+        
+        // Submit Edit Location Form
+        $(document).on('submit', '#editLocationForm', function (e) {
+            e.preventDefault();
+            $('.validation-error-block').remove();
+            $(".submitBtn").attr('disabled', true);
+            var formData = new FormData(this);
+
+            var url = $(this).data('action');
+
+            $.ajax({
+                type: 'post',
+                url: url,
+                dataType: 'json',
+                contentType: false,
+                processData: false,
+                data: formData,
+                success: function (response) {
+                    if(response.success) {
+                        $('#editLocationModal').modal('hide');
+                        fireSuccessSwal('Success',response.message);
+                        $('#location-table').DataTable().ajax.reload(null, false);
+                    }
+                },
+                error: function (response) {
+                    $(".submitBtn").attr('disabled', false);
+                    if(response.error_type == 'something_error'){
+                        fireErrorSwal('Error',response.message);
+                    } else {                    
+                        var errorLabelTitle = '';
+                        $.each(response.responseJSON.errors, function (key, item) {
+                            errorLabelTitle = '<span class="validation-error-block">'+item+'</sapn>';
+                            if(key == 'resume' || key == 'gender'){
+                                $('#'+key).html(errorLabelTitle);
+                            }
+                            else{
+                                $(errorLabelTitle).insertAfter("input[name='"+key+"']");
+                            }                    
+                        });
+                    }
+                },
+                complete: function(res){
+                    $(".submitBtn").attr('disabled', false);
+                }
+            });
+        });
+    @endcan
+
+    @can('location_delete')
+        $(document).on("click",".deleteLocationBtn", function() {
+            var url = $(this).data('href');
+            Swal.fire({
+            title: "{{ trans('global.areYouSure') }}",
+            text: "{{ trans('global.onceClickedRecordDeleted') }}",
+            icon: "warning",
+            showDenyButton: true,  
+            //   showCancelButton: true,  
+            confirmButtonText: "Yes, I am sure",  
+            denyButtonText: "No, cancel it!",
+            })
+            .then(function(result) {
+                if (result.isConfirmed) {  
+                    $.ajax({
+                        type: 'DELETE',
+                        url: url,
+                        dataType: 'json',
+                        data: { _token: "{{ csrf_token() }}" },
+                        success: function (response) {
+                            if(response.success) {
+                                fireSuccessSwal('Success',response.message);
+                                $('#location-table').DataTable().ajax.reload(null, false);
+                            }
+                            else {
+                                fireErrorSwal('Error',response.message);
+                            }
+                        },
+                        error: function(res){
+                            fireErrorSwal('Error',res.message);
                         }
-                        else{
-                            $(errorLabelTitle).insertAfter("input[name='"+key+"']");
-                        }                    
                     });
                 }
-            },
-            complete: function(res){
-                $(".submitBtn").attr('disabled', false);
-            }
-        });                    
-    });
-
-    // Edit Location Modal
-    $(document).on("click",".editLocationBtn", function() {
-        // $('#pageloader').css('display', 'flex');
-        var url = $(this).data('href');
-        $.ajax({
-            type: 'get',
-            url: url,
-            dataType: 'json',
-            //data: formData,
-            success: function (response) {
-                // $('#pageloader').css('display', 'none');
-                if(response.success) {
-                    $('.popup_render_div').html(response.htmlView);
-                    $('#editLocationModal').modal('show');
-                }
-            }
+            });
         });
-    });
-    
-    // Submit Edit Location Form
-    $(document).on('submit', '#editLocationForm', function (e) {
-        e.preventDefault();
-        $('.validation-error-block').remove();
-        $(".submitBtn").attr('disabled', true);
-        var formData = new FormData(this);
 
-        var url = $(this).data('action');
-
-        $.ajax({
-            type: 'post',
-            url: url,
-            dataType: 'json',
-            contentType: false,
-            processData: false,
-            data: formData,
-            success: function (response) {
-                if(response.success) {
-                    $('#editLocationModal').modal('hide');
-                    fireSuccessSwal('Success',response.message);
-                    $('#location-table').DataTable().ajax.reload(null, false);
-                }
-            },
-            error: function (response) {
-                $(".submitBtn").attr('disabled', false);
-                if(response.error_type == 'something_error'){
-                    fireErrorSwal('Error',response.message);
-                } else {                    
-                    var errorLabelTitle = '';
-                    $.each(response.responseJSON.errors, function (key, item) {
-                        errorLabelTitle = '<span class="validation-error-block">'+item+'</sapn>';
-                        if(key == 'resume' || key == 'gender'){
-                            $('#'+key).html(errorLabelTitle);
+        $(document).on('click', '#deleteAllLocation', function(e){
+            e.preventDefault();
+            var t = $(this);
+            var selectedIds = [];
+            $('.dt_cb:checked').each(function() {
+                selectedIds.push($(this).data('id'));
+            });
+            if(selectedIds.length == 0){
+                fireWarningSwal('Warning', "{{ trans('messages.warning_select_record') }}");
+                return false;
+            }
+            Swal.fire({
+            title: "{{ trans('global.areYouSure') }}",
+            text: "{{ trans('global.onceClickedRecordDeleted') }}",
+            icon: "warning",
+            showDenyButton: true,  
+            //   showCancelButton: true,  
+            confirmButtonText: "Yes, I am sure",  
+            denyButtonText: "No, cancel it!",
+            })
+            .then(function(result) {
+                if (result.isConfirmed) {       
+                    console.log(selectedIds);   
+                    $.ajax({
+                        url: "{{route('locations.massDestroy')}}",
+                        type: "POST",
+                        data: { 
+                            ids: selectedIds,
+                            _token: "{{ csrf_token() }}",
+                        },
+                        dataType: 'json',
+                        success: function (response) {
+                            if(response.success) {
+                                fireSuccessSwal('Success',response.message);
+                                $('#location-table').DataTable().ajax.reload(null, false);
+                            }
+                            else {
+                                fireErrorSwal('Error',response.message);
+                            }
+                        },
+                        error: function(res){
+                            fireErrorSwal('Error',res.message);
                         }
-                        else{
-                            $(errorLabelTitle).insertAfter("input[name='"+key+"']");
-                        }                    
-                    });
+                    })
                 }
-            },
-            complete: function(res){
-                $(".submitBtn").attr('disabled', false);
-            }
-        });
-    });
-    
-
-    $(document).on("click",".deleteLocationBtn", function() {
-        var url = $(this).data('href');
-        Swal.fire({
-          title: "{{ trans('global.areYouSure') }}",
-          text: "{{ trans('global.onceClickedRecordDeleted') }}",
-          icon: "warning",
-          showDenyButton: true,  
-        //   showCancelButton: true,  
-          confirmButtonText: "Yes, I am sure",  
-          denyButtonText: "No, cancel it!",
+            });
         })
-        .then(function(result) {
-            if (result.isConfirmed) {  
-                $.ajax({
-                    type: 'DELETE',
-                    url: url,
-                    dataType: 'json',
-                    data: { _token: "{{ csrf_token() }}" },
-                    success: function (response) {
-                        if(response.success) {
-                            fireSuccessSwal('Success',response.message);
-                            $('#location-table').DataTable().ajax.reload(null, false);
-                        }
-                        else {
-                            fireErrorSwal('Error',response.message);
-                        }
-                    },
-                    error: function(res){
-                        fireErrorSwal('Error',res.message);
-                    }
-                });
-            }
-        });
-    });
-
-    $(document).on('click', '#deleteAllLocation', function(e){
-        e.preventDefault();
-        var t = $(this);
-        var selectedIds = [];
-        $('.dt_cb:checked').each(function() {
-            selectedIds.push($(this).data('id'));
-        });
-        if(selectedIds.length == 0){
-            fireWarningSwal('Warning', "{{ trans('messages.warning_select_record') }}");
-            return false;
-        }
-        Swal.fire({
-          title: "{{ trans('global.areYouSure') }}",
-          text: "{{ trans('global.onceClickedRecordDeleted') }}",
-          icon: "warning",
-          showDenyButton: true,  
-        //   showCancelButton: true,  
-          confirmButtonText: "Yes, I am sure",  
-          denyButtonText: "No, cancel it!",
-        })
-        .then(function(result) {
-            if (result.isConfirmed) {       
-                console.log(selectedIds);   
-                $.ajax({
-                    url: "{{route('locations.massDestroy')}}",
-                    type: "POST",
-                    data: { 
-                        ids: selectedIds,
-                        _token: "{{ csrf_token() }}",
-                    },
-                    dataType: 'json',
-                    success: function (response) {
-                        if(response.success) {
-                            fireSuccessSwal('Success',response.message);
-                            $('#location-table').DataTable().ajax.reload(null, false);
-                        }
-                        else {
-                            fireErrorSwal('Error',response.message);
-                        }
-                    },
-                    error: function(res){
-                        fireErrorSwal('Error',res.message);
-                    }
-                })
-            }
-        });
-    })
+    @endcan
 </script>
 
 @endsection
